Payments Post() being tested for Location header
To accomplish this, Route Configuration was added to Api.  WebHost and Tests uses Api's routes so Tests don't have to take a dependency on the host.  A Web Host could choose to override this, but Api controllers have a dependency on stringly typed route names to use CreatedAtRoute<T>().  Not sure it is worth the effort to decouple this.  Web API Controllers and Routes just might be coupled.