Assembly discovery problem in some tests fixed
The Acceptance tests (formerly Api.WebHost tests) were failing because
the Api assembly was not being discovered by the
DefaultAssembliesResolver. Changed the concept from testing the
integration of api and web host to a concept of the tests being
Acceptance Tests, which still may be more useful to some people.